refactor: remove LottoResult and add PrizeCalculator
