--- src/main/java/lotto/service/LottoResultCalculator.java
@@ -1,4 +1,0 @@
-package lotto.service;
-
-public class LottoResultCalculator {
-}
