--- src/test/java/lotto/domain/LottoResultTest.java
@@ -1,20 +1,0 @@
-package lotto.domain;
-
-class LottoResultTest {
-//    @Test
-//    @DisplayName("구입 금액과 당첨 결과를 파라미터로 LottoResult 객체가 생성되어야 한다")
-//    void create() {
-//        // given
-//        final LottoPayment payment = new LottoPayment("10000");
-//        final LottoTickets tickets = new LottoTickets(payment);
-//        final LottoNumbers winningNumbers = new LottoNumbers(Arrays.asList(1, 2, 3, 4, 5, 6));
-//
-//        // when
-//        final LottoResult lottoResult = new LottoResult(payment, tickets, winningNumbers);
-//        lottoResult.printResult();
-//
-//        // then
-//        assertThat(lottoResult).isInstanceOf(LottoResult.class);
-//        assertThat(lottoResult).isEqualTo(new LottoResult(payment, tickets, winningNumbers));
-//    }
-}
